+++ app/src/main/java/woowacourse/shopping/data/repository/ProductRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.repository
 
 import woowacourse.shopping.model.Product
